Moved GUI related code to separate module
